pyproject.toml
[tool.poetry]
name = "cgfs"
version = "0.1.0"
description = ""
authors = ["Colin Murphy <cjm@cathect.net>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
